+++ shared/src/commonMain/kotlin/business/interactors/main/LikeInteractor.kt
@@ -16,7 +16,7 @@ class LikeInteractor(
 ) {
 
 
-    fun execute(id: Int): Flow<DataState<Boolean>> = flow {
+    fun execute(id: Long): Flow<DataState<Boolean>> = flow {
 
         try {
